Added population df cleanup as a part of the generation process
